+++ ios/flutter_upgrader_plus/Package.swift
@@ -1,0 +1,25 @@
+// swift-tools-version: 5.9
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let package = Package(
+    name: "flutter_upgrader_plus",
+    platforms: [
+        .iOS("12.0")
+    ],
+    products: [
+        .library(name: "flutter-upgrader-plus", targets: ["flutter_upgrader_plus"])
+    ],
+    dependencies: [
+        .package(name: "FlutterFramework", path: "../FlutterFramework")
+    ],
+    targets: [
+        .target(
+            name: "flutter_upgrader_plus",
+            dependencies: [
+                .product(name: "FlutterFramework", package: "FlutterFramework")
+            ]
+        )
+    ]
+)